refactor: Move new classes to top-level model folder
