+++ event-handlers/src/main/java/no/sikt/nva/nvi/events/cristin/CristinNviReport.java
@@ -8,7 +8,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import java.util.List;
 import no.sikt.nva.nvi.common.db.CandidateDao.DbLevel;
-import no.sikt.nva.nvi.common.service.model.PublicationDate;
+import no.sikt.nva.nvi.common.model.PublicationDate;
 import no.unit.nva.commons.json.JsonSerializable;
 
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
